--- modulo1_sesion3/main3.js
@@ -1,3 +1,0 @@
-import {mensajeAlerta} from './mensajeAlerta';
-
-mensajeAlerta("hola alumno");
