--- modulo1_sesion3/mensajeAlerta.js
@@ -1,3 +1,0 @@
-export const mensajeAlerta = (mensaje) => {
-    console.log(mensaje);
-}
